Fix for the Lapack illegal parameter issue
The code that meant to avoid recompression (ie premature optimization)
also happened to fix the issue with h_lu 5 and 6. It caused the issue
with Lapack as well, but that is fixed in this commit.
The issue was: In Hierarchical*Hierarchical, it was attempted to
compress a (16x1) dense matrix (ie a vector) which means that the rank
is bigger than one of the dimensions and thus illegal.
Checking whether the result is a vector before possibly recompressing
fixed both the Lapack issue and the problem with h_lu cases 5 and 6.

For deeper recursions, LowRank * Hierarchical still needs to be
implemented.