BLR LU with uniform basis implemented in python
Error in approximation is higher than in original version, but end
result has the same precision. One reason for this might be that the
blocks used to calculate the basis are currently not the ones closest to
the diagonal, but simply the first encountered that can be used to
calculate new parts of the basis.
A next step would be to fix this. Afterwards, the harder step to a
hierarchical basis needs to be taken.